Inline enkf_main_fs_manager
